requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=10.0.0
tifffile>=2023.0.0
matplotlib>=3.7.0
tensorboard>=2.13.0
pyyaml>=6.0
tqdm>=4.65.0

